Made Mandrill email notifier save the data to a file first.
This means the curl call functions as it does with Mailgun, meaning we don't get the "args too long" issue!

Also removed the -v on create and check, we only want the stats!

I feel like we should deal with the logs better, maybe a single log file that gets rotated rather than a new one each day?